repos/banuvarsha10_fin-rag_genai/finsage/requirements.txt
streamlit
sentence-transformers
faiss-cpu
pypdf
groq
numpy
pandas
matplotlib
scipy>=1.11